# btcec: Optimize field inversion.