This uses the addition chain from [Brian Smith's website](https://briansmith.org/ecc-inversion-addition-chains-01#secp256k1_field_inversion) to speed up
the inversion algorithm for btcec/field.go.